v1 rule files should replace v2 rule files instead of merge.
