+++ setup/gcp/run_forseti_installer.py
@@ -19,9 +19,30 @@ This has been tested with python 2.7.
 
 import argparse
 import datetime
+import site
 
-from installer.forseti_server_installer import ForsetiServerInstaller
-from installer.forseti_client_installer import ForsetiClientInstaller
+import pip
+
+INSTALLER_REQUIRED_PACKAGES = [
+    'ruamel.yaml'
+]
+
+
+def install(package_name):
+    """Install package.
+
+    Args:
+        package_name (str): Name of the package to install.
+    """
+    pip.main(['install', package_name, '--user'])
+
+
+def install_required_packages():
+    """Install required packages."""
+    installed_pkgs = [pkg.key for pkg in pip.get_installed_distributions()]
+    for package in INSTALLER_REQUIRED_PACKAGES:
+        if package not in installed_pkgs:
+            install(package)
 
 
 def run():
@@ -67,6 +88,10 @@ def run():
     # Set the current date time stamp
     args['datetimestamp'] = datetime.datetime.now().strftime('%Y%m%d%H%M%S')
 
+    # import installers
+    from installer.forseti_server_installer import ForsetiServerInstaller
+    from installer.forseti_client_installer import ForsetiClientInstaller
+
     if not args.get('type'):
         # If the user didn't specify a type, install both server and client
         ForsetiServerInstaller(**args).run_setup()
@@ -83,4 +108,9 @@ def run():
     forseti_setup.run_setup()
 
 if __name__ == '__main__':
+    # We need to install all the required packages before importing our modules
+
+    # Installing required packages
+    install_required_packages()
+    site.main() # Load up the package
     run()
