+++ setup/installer.py
@@ -16,41 +16,8 @@
 
 A stub to call gcp/run_forseti_installer.py which installs into GCP.
 """
-
-import site
-
-import pip
-
-
-INSTALLER_REQUIRED_PACKAGES = [
-    'ruamel.yaml'
-]
-
-def install(package_name):
-    """Install package.
-
-    Args:
-        package_name (str): Name of the package to install.
-    """
-    pip.main(['install', package_name, '--user'])
-
-
-def install_required_packages():
-    """Install required packages."""
-    installed_pkgs = [pkg.key for pkg in pip.get_installed_distributions()]
-    for package in INSTALLER_REQUIRED_PACKAGES:
-        if package not in installed_pkgs:
-            install(package)
-
+from gcp import run_forseti_installer
 
 
 if __name__ == '__main__':
-    # We need to install all the required packages before importing our modules
-
-    # Installing required packages
-    install_required_packages()
-    site.main() # Load up the package
-
-    # Importing our own modules
-    from gcp import run_forseti_installer
     run_forseti_installer.run()
